+++ src/Products/PloneMeeting/vocabularies.py
@@ -26,8 +26,8 @@ from collective.eeafaceted.dashboard.vocabulary import DashboardCollectionsVocab
 from collective.eeafaceted.z3ctable.columns import EMPTY_STRING
 from collective.iconifiedcategory.config import get_sort_categorized_tab
 from collective.iconifiedcategory.utils import get_categorized_elements
-from collective.iconifiedcategory.utils import get_category_object
 from collective.iconifiedcategory.utils import get_category_icon_url
+from collective.iconifiedcategory.utils import get_category_object
 from collective.iconifiedcategory.utils import get_config_root
 from collective.iconifiedcategory.utils import get_group
 from collective.iconifiedcategory.utils import render_filesize
@@ -3263,11 +3263,23 @@ class WorkflowAdaptationsVocabulary(object):
         """Received "context" is a MeetingConfig."""
         terms = []
         for adaptation in context.wfAdaptations:
+            # generate a WFA by MeetingConfig.powerObservers in addition to the base one
+            if adaptation == 'hide_decisions_when_under_writing':
+                tool = api.portal.get_tool('portal_plonemeeting')
+                cfg = tool.getMeetingConfig(context)
+                for po in cfg.getPowerObservers():
+                    term_id = 'hide_decisions_when_under_writing__po__{0}'.format(po['row_id'])
+                    title = translate(
+                        'wa_hide_decisions_when_under_writing_excepted_po',
+                        domain='PloneMeeting',
+                        mapping={'po': safe_unicode(po['label'])},
+                        context=context.REQUEST)
+                    terms.append(SimpleTerm(term_id, term_id, title))
             # back transitions from presented to every available item validation
             # states defined in MeetingConfig.itemWFValidationLevels
             if adaptation == 'presented_item_back_to_validation_state':
                 for item_validation_level in context.getItemWFValidationLevels(only_enabled=True):
-                    adaptation_id = 'presented_item_back_to_{0}'.format(item_validation_level['state'])
+                    term_id = 'presented_item_back_to_{0}'.format(item_validation_level['state'])
                     translated_item_validation_state = translate(
                         safe_unicode(item_validation_level['state_title']),
                         domain='plone',
@@ -3279,8 +3291,8 @@ class WorkflowAdaptationsVocabulary(object):
                         context=context.REQUEST,
                         default=u'Item back to presented from validation state "{0}"'.format(
                             translated_item_validation_state))
-                    title = title + " ({0})".format(adaptation_id)
-                    terms.append(SimpleTerm(adaptation_id, adaptation_id, title))
+                    title = title + " ({0})".format(term_id)
+                    terms.append(SimpleTerm(term_id, term_id, title))
             else:
                 title = translate('wa_%s' % adaptation, domain='PloneMeeting', context=context.REQUEST)
                 title = title + " ({0})".format(adaptation)